refactor(update): isolate release candidate selection
