+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
   "schema_version": 1,
   "maximum_collected_cases": 16300,
-  "maximum_test_source_lines": 352489
+  "maximum_test_source_lines": 352495
 }
